amebazii

Module keys

Default keys used to sign and verify images/ partition tables.

Structs

ConstKeyPair

Represents a pair of cryptographic keys: a private key and a public key.

Constants

APP_DEFAULT_USER_KEY2

Default second key used in the application subimage.

BOOT_DEFAULT_USER_KEY1

Default first key used in the boot image.

DEFAULT_IV

The default Initialization Vector (IV) used for encryption/decryption operations.

DEFAULT_VALID_PATTERN

FLASH_PATTERN

Default flash calibration pattern used in flash

HASH_KEY

The default hash key used to generate signatures for the partition table.

KEY_PAIR_000

Default cryptographic key pair for general use in encryption/decryption images.

KEY_PAIR_001

Default hash key pair used for hashing operations (not seen).

KEY_PAIR_002

Unused key pair; not seen

KEY_PAIR_003

Default key pair used for firmware signature generation (used in OTA operations).

XIP_IV

Default IV used for encrypting/decrypting XIP section data.

XIP_KEY

Default AES encryption key used for encrypting/decrypting XIP section data.