amebazii
Module map

Structs

AddressRange

Represents a memory address range, defined by a starting address and an ending address.

Constants

DTCM_RAM

Internal RAM for program data and text (Data and Text are loaded into DTCM RAM).

EXTENSION_RAM

Extension RAM for heap, used as dynamic memory.

PSRAM

External PSRAM for storing text, read-only data (RODATA), and data sections.

RAM_FUN_TABLE

RAM functions entry table, storing function entries in RAM.

RAM_IMG_SIGN

RAM image signature, used for storing the image signature in RAM.

VECTORS_RAM

The vector table, it must start with 256 bytes aligned address.

XIP_FLASH_C

XIP Chiper section, where TEXT and RODATA sections can be encrypted.

XIP_FLASH_P

XIP Plaintext section, where RODATA is not encrypted.