amebazii::types::enums

Enum SectionType

#[repr(u8)]
pub enum SectionType {
    DTCM = 128,
    ITCM = 129,
    SRAM = 130,
    PSRAM = 131,
    LPDDR = 132,
    XIP = 133,
}

Enum representing different section types in memory.

This enum defines the types of memory sections that can exist, each represented by a specific identifier (u8 value). These sections correspond to various types of memory regions, such as data cache memory (DTCM), instruction cache memory (ITCM), and other specialized memory regions.

Variants

DTCM: Data tightly coupled memory (0x80).

ITCM: Instruction tightly coupled memory (0x81).

SRAM: Static RAM (0x82).

PSRAM: Pseudo-static RAM (0x83).

LPDDR: Low power DDR memory (0x84).

XIP: Execute-In-Place memory (0x85), containing raw binary with compiled code.

The XIP variant refers to memory regions that can execute code directly from the memory, without the need to copy the code into RAM.

Example

use amebazii::types::enums::SectionType;

let section = SectionType::try_from(0x80).unwrap();
assert_eq!(section, SectionType::DTCM); // Successfully converts to DTCM.

Error Handling

If the provided value doesnโt correspond to a known section type, an error (Error::UnknownSectionType) will be returned.

Variants

DTCM = 128

ITCM = 129

SRAM = 130

PSRAM = 131

LPDDR = 132

XIP = 133

Execute-In-Place (XIP) contains the raw binary with all compiled code.