amebazii::types::header
Struct ImageHeader

pub struct ImageHeader {
    pub segment_size: u32,
    pub next_offset: Option<u32>,
    pub img_type: ImageType,
    pub is_encrypt: bool,
    pub serial: u32,
    pub user_key1: DataType<32>,
    pub user_key2: DataType<32>,
}

Generic image header.

This struct contains metadata for an image, such as the segment size, offset to the next image header, the image type, encryption flag, and user keys.

Fields
segment_size: u32

The size of the image segment in bytes.

This field specifies the size of the image’s data segment, which can be used to determine how much data is associated with the current image header.

next_offset: Option<u32>

Offset to the next image header.

If there is no next image header, this field is set to None. Otherwise, it holds the byte offset to the next image header.

img_type: ImageType

The type of the image.

This field stores the image type, which can represent different image types such as boot images, partition tables, etc.

is_encrypt: bool

Flag indicating whether the image is encrypted.

This boolean flag indicates whether the image is encrypted (true) or not (false).

serial: u32

The serial number associated with the image. (version number)

This field stores the image’s serial number. It is initialized to 0xFFFF_FFFF by default.

user_key1: DataType<32>

User key 1, used for encryption

user_key2: DataType<32>

User key 2, used for encryption

Implementations

impl ImageHeader

pub fn is_key1_valid(&self) -> bool

Checks if the first user key (user_key1) is valid.

Returns

true if user_key1 is valid (i.e., not all bytes are 0xFF).

false if user_key1 is invalid (i.e., all bytes are 0xFF).

pub fn is_key2_valid(&self) -> bool

Checks if the second user key (user_key2) is valid.

Returns

true if user_key2 is valid (i.e., not all bytes are 0xFF).

false if user_key2 is invalid (i.e., all bytes are 0xFF).

pub fn has_next(&self) -> bool

Checks if there is a next image header.

The next_offset field indicates the offset to the next image header. If the value is 0xFFFF_FFFF, it means there is no next image. This method returns true if the next_offset is not 0xFFFF_FFFF, indicating there is a next image header.

Returns

true if there is a next image (i.e., next_offset is not 0xFFFF_FFFF).

false if there is no next image (i.e., next_offset is 0xFFFF_FFFF).

pub fn get_user_key1(&self) -> DataRefType<'_, 32>

Gets the first user key (user_key1).

Returns

A reference to the user_key1 array (32 bytes).

pub fn get_user_key2(&self) -> DataRefType<'_, 32>

Gets the second user key (user_key2).

Returns

A reference to the user_key2 array (32 bytes).

pub fn set_user_key1(&mut self, key: DataType<32>)

Sets the first user key (user_key1) in the image header.

Arguments

key - A DataType<32> representing the new value for user_key1.

pub fn set_user_key2(&mut self, key: DataType<32>)

Sets the second user key (user_key2) in the image header.

Arguments

key - A DataType<32> representing the new value for user_key2.