amebazii::types::header

Struct KeyBlock

pub struct KeyBlock { /* private fields */ }

A struct representing the key block with two public keys:

an encryption public key and

a hash public key.

Implementations

impl KeyBlock

pub fn is_enc_pubkey_valid(&self) -> bool

Checks if the encryption public key is valid.

Returns

true if the encryption public key is valid.

false if the encryption public key is invalid.

pub fn is_hash_pubkey_valid(&self) -> bool

Checks if the hash public key is valid.

Returns

true if the hash public key is valid.

false if the hash public key is invalid.

pub fn get_enc_pubkey(&self) -> &[u8; 32]

Retrieves the encryption public key.

Returns

A reference to the encryption public key (32 bytes).

pub fn get_hash_pubkey(&self) -> &[u8; 32]

Retrieves the hash public key.

Returns

A reference to the hash public key (32 bytes).

pub fn get_enc_pubkey_mut(&mut self) -> &mut [u8; 32]

Retrieves a mutable reference to the encryption public key.

Returns

A mutable reference to the encryption public key (32 bytes).

pub fn get_hash_pubkey_mut(&mut self) -> &mut [u8; 32]

Retrieves a mutable reference to the hash public key.

Returns

A mutable reference to the hash public key (32 bytes).