amebazii::types::image

Enum EncryptedOr

pub enum EncryptedOr<T> {
    Encrypted(Vec<u8>),
    Plain(T),
}

A generic enum representing either encrypted or plain data.

The EncryptedOr enum is used to differentiate between encrypted data and unencrypted (plain) data. It allows to store either type of data in the same structure while providing methods to safely access or mutate the contents, depending on whether the data is encrypted or not.

Variants

Encrypted(Vec<u8>)

Contains encrypted data as a vector of bytes.

Plain(T)

Contains plain, unencrypted data of type T.

Implementations

impl<T> EncryptedOr<T>

pub fn is_encrypted(&self) -> bool

Returns true if the data is encrypted.

This method allows checking if the current instance of EncryptedOr contains encrypted data.

pub fn is_plain(&self) -> bool

Returns true if the data is plain.

This method allows checking if the current instance of EncryptedOr contains plain (unencrypted) data.

pub fn unwrap(self) -> T