amebazii::types::image::pt

Struct Record

pub struct Record {
    pub start_addr: u32,
    pub length: u32,
    pub part_type: PartitionType,
    pub dbg_skip: bool,
    /* private fields */
}

Represents a firmware partition record.

A Record struct encapsulates information about a partition within a firmware image. This record stores a hash key for the partition, which will be later used to verify the signatures of the partition contents.

Layout (64 bytes):

         +---------------+---+---+---+---+---+---+---+----------+--------------+----+----+----+----+----+----+
         | 0             | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8        | 9            | 10 | 11 | 12 | 13 | 14 | 15 |
+========+===============+===+===+===+===+===+===+===+==========+==============+====+====+====+====+====+====+
| 0x00   |      start_addr: u32      |  length: u32  | type: u8 | dbg_skip: u8 |                             |
+--------+---------------------------+---------------+----------+--------------+-----------------------------+
| 0x10   | key_valid: u8 |                                                                                   |
+--------+---------------+-----------------------------------------------------------------------------------+
| 0x20   |                                        hash_key: bytes[32]                                        |
+--------+---------------------------------------------------------------------------------------------------+

Fields
start_addr: u32

The starting address of the partition in the firmware image (4 bytes).

length: u32

The length of the partition in bytes (4 bytes).

part_type: PartitionType

The partition type (1 byte). This is an enum (PartitionType).

dbg_skip: bool

A flag that indicates whether debugging should be skipped for this partition (1 byte). true means skip debugging, false means do not skip.

Implementations

impl Record

pub fn hash_key_valid(&self) -> bool

Checks whether the hash_key is valid.

Returns:

true if the hash_key is valid (non-None and passes the validation check).

false otherwise (e.g., if the key is None or invalid).

pub fn get_hash_key(&self) -> DataRefType<'_, 32>

Returns a reference to the hash_key.

pub fn set_hash_key(&mut self, key: DataType<32>)

Sets the hash_key to a new value.

Arguments:

key: A reference to a 32-byte array slice that represents the new hash_key. If None is passed, the hash_key will be cleared.