amebazii::types::sysctrl
Struct SystemData

pub struct SystemData {
    pub ota2_addr: Option<u32>,
    pub ota2_size: Option<u32>,
    pub old_img_trap: ForceOldImage,
    pub spi_cfg: SpiConfig,
    pub flash_info: FlashInfo,
    pub ulog_baud: u32,
    /* private fields */
}

Represents system data related to the device, including configuration for OTA, SPI, and other hardware settings.

The SystemData struct holds various fields representing the system’s configuration and settings:

OTA: Configurations for over-the-air (OTA) updates, including address and size of OTA2.

Force Old Image: Settings related to forcing the usage of an older firmware image.

SPI Configuration: Configuration for the Serial Peripheral Interface (SPI) bus.

Flash Information: Information about the flash memory, including ID and size.

Logging Baud Rate: Baud rate for UART logging.

Calibration Data: SPI calibration data and Bluetooth parameters.

This struct is designed to be serialized and deserialized from a stream, allowing for easy reading and writing of system settings, especially in embedded systems or firmware configuration.

Fields
ota2_addr: Option<u32>

Address of the OTA2 image, or None if OTA1 is active.

ota2_size: Option<u32>

Size of the OTA2 image, or None if OTA1 is active.

old_img_trap: ForceOldImage

Configuration for forcing the usage of an older image, including the pin, port, and active status.

spi_cfg: SpiConfig

Configuration for the SPI bus, including IO mode and speed.

flash_info: FlashInfo

Information about the flash memory, including ID and size.

ulog_baud: u32

Baud rate for the UART logging interface.

Implementations

impl SystemData

pub fn get_bt_paramdata(&self) -> DataRefType<'_, 32>

Retrieves the Bluetooth parameter data as a reference.

Returns:

A reference to the Bluetooth parameter data (if available).

pub fn get_spic_calibcfg(&self) -> DataRefType<'_, 48>

Retrieves the SPI calibration configuration as a reference.

Returns:

A reference to the SPI calibration data (if available).

pub fn set_pt_paramdata(&mut self, data: DataType<32>)

Sets the Bluetooth parameter data.

Parameters:

data: The new Bluetooth parameter data to set.

pub fn set_spic_calibcfg(&mut self, data: DataType<48>)

Sets the SPI calibration configuration.

Parameters:

data: The new SPI calibration data to set.