Crate amebazii

See the CLI documentation.

Modules

conf

doc

Documentation

error

keys

Default keys used to sign and verify images/ partition tables.

map

types

util

Utility functions for creating signatures and hashes.

Macros

expect_length

is_valid_data

Checks if the given data is valid by ensuring none of the bytes are equal to 0xFF.

read_padding

Read (skip) padding bytes in a reader.

read_valid_data

Reads valid data from the reader into the target, ensuring that the data does not contain any 0xFF bytes.

write_aligned

Writes padding to a binary stream to ensure that the next write operation aligns to a specified size.

write_data

Writes data to a stream.

write_padding

write_padding! - A macro to write padding bytes to a writer.