amebazii

Module types

Re-exports

pub use flash::Flash;

pub use flash::Partition;

pub use fst::FST;

pub use header::EntryHeader;

pub use header::ImageHeader;

pub use header::KeyBlock;

pub use header::SectionHeader;

pub use section::Section;

pub use sysctrl::FlashInfo;

pub use sysctrl::ForceOldImage;

pub use sysctrl::SpiConfig;

pub use sysctrl::SystemData;

pub use enums::*;

pub use image::*;

Modules

enums

flash

fst

header

image

section

sysctrl

Traits

BinarySize

A trait for types that can provide their binary size.

FromStream

A trait for types that can be deserialized from a stream.

ToStream

A trait for types that can be serialized to a stream.

Functions

from_stream

Reads a type from a stream.

key_from_hex

Converts a hexadecimal string into a DataType array.

key_to_hex

Converts a DataType array into a hexadecimal string.

to_bytes

Serializes an object into a vector of bytes.

to_bytes_with_capacity

Serializes an object into a vector of bytes with an optimized capacity.

transfer_to

Transfers a type’s data to a stream.

Type Aliases

DataRefType

DataRefType is a type alias for an optional reference to a fixed-size array of u8 bytes.

DataType

DataType is a type alias for an optional fixed-size array of u8 bytes.