amebazii::types

Module image

Re-exports

pub use boot::BootImage;

pub use ota::OTAImage;

pub use ota::SubImage;

pub use pt::PartTab;

pub use pt::PartitionTableImage;

pub use pt::Record;

pub use pt::TrapConfig;

Modules

boot

ota

pt

Enums

EncryptedOr

A generic enum representing either encrypted or plain data.

Traits

AsImage

A trait that provides common functionality for image-like objects, such as computing and setting the segment size and signature.

Functions

build_default_signature

Builds the signature for a given image.

build_segment_size

Builds the segment size for a given image.

set_default_segment_size

Sets the segment size for a given image.

set_default_signature

Sets the signature for a given image.

Type Aliases

RawImage