amebazii
Module util

Utility functions for creating signatures and hashes.

Functions

hmac_md5

Computes an HMAC-MD5 signature for the provided key and data.

hmac_sha256

Computes an HMAC-SHA256 signature for the provided key and data.

md5

Computes an MD5 hash of the provided data.

sha256

Computes a SHA-256 hash of the provided data.

skip_aligned

Skips bytes in the provided reader to ensure that the next read operation aligns with the specified alignment.

write_fill

Writes the specified byte fill repeatedly to the writer for the given length.